requirements.txt.txt
torch
torchvision
torchaudio
transformers==4.36.2
tokenizers==0.15.0
pillow==9.5.0
matplotlib==3.7.2
numpy==1.24.4
scipy==1.10.1
tqdm==4.66.1
pytorch-pretrained-biggan==0.1.1